Action configuration became optional
